+++ pySIMsalabim/experiments/EQE.py
@@ -1,6 +1,4 @@
-"""Perform EQE simulations
-## Version: 1 (16-10-2024) ## Internal version number for this script. Refer to the changelog to review changes ## 
-"""
+"""Perform EQE simulations"""
 
 ######### Package Imports #########################################################################