+++ pySIMsalabim/experiments/JV_steady_state.py
@@ -1,6 +1,4 @@
-"""Perform steady-state JV simulations
-## Version: 1 (16-10-2024) ## Internal version number for this script. Refer to the changelog to review changes ## 
-"""
+"""Perform steady-state JV simulations"""
 
 ######### Package Imports #########################################################################