+++ pySIMsalabim/experiments/imps.py
@@ -1,6 +1,4 @@
-"""Perform Intensity-Modulated Photocurrent Spectroscopy (IMPS) simulations
-## Version: 1 (16-10-2024) ## Internal version number for this script. Refer to the changelog to review changes ## 
-"""
+"""Perform Intensity-Modulated Photocurrent Spectroscopy (IMPS) simulations"""
 ######### Package Imports #########################################################################
 
 import os, sys